+++ changelog.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file, starting
 with 1.0.0beta.
 
+## [1.2-network-partners] Unmerged (branched from `1.2-dev`)
+
+- Improve: Page template padding, to allow for full-width page elements.
+
 ## [1.2-jobs] - Unmerged (branched from `1.2-dev`)
 
 - Add: Single Job route